Fix salt alignment in tacacs-plus format

+++ src/tacacs_plus_fmt_plug.c
@@ -72,7 +72,7 @@ john_register_one(&fmt_tacacsplus);
 #define BINARY_SIZE             0
 #define BINARY_ALIGN            1
 #define SALT_SIZE               sizeof(struct custom_salt)
-#define SALT_ALIGN              sizeof(uint32_t)
+#define SALT_ALIGN              sizeof(uint64_t)
 #define MIN_KEYS_PER_CRYPT      1
 #define MAX_KEYS_PER_CRYPT      64  // tuned on i5-6500 CPU
